Move model from experimental to demos and add demo CI
