+++ models/demos/yolov8s/README.md
@@ -23,7 +23,7 @@ Resource link - [source](https://github.com/ultralytics/ultralytics/blob/main/ul
 
 ### Details
 
-- The entry point to the yolov8s is located at:`models/experimental/yolov8s/tt/ttnn_yolov8s.py`
+- The entry point to the yolov8s is located at:`models/demos/yolov8s/tt/ttnn_yolov8s.py`
 - Batch Size :1
 - Supported Input Resolution - (640,640) (Height,Width)
 
@@ -42,5 +42,5 @@ pytest --disable-warnings tests/ttnn/integration_tests/yolov8s/test_yolov8s.py::
 Use the following command to run the performant Model with Trace+2CQs:
 
 ```
-pytest --disable-warnings models/experimental/yolov8s/tests/test_e2e_performant.py
+pytest --disable-warnings models/demos/yolov8s/tests/test_e2e_performant.py
 ```
