+++ models/demos/yolov8s/tests/test_e2e_performant.py
@@ -8,8 +8,7 @@ import pytest
 import torch
 from loguru import logger
 
-import ttnn
-from models.experimental.yolov8s.tests.yolov8s_e2e_performant import Yolov8sTrace2CQ
+from models.demos.yolov8s.tests.yolov8s_e2e_performant import Yolov8sTrace2CQ
 from models.utility_functions import run_for_wormhole_b0
 
 
@@ -37,13 +36,12 @@ def test_run_yolov8s_trace_2cqs_inference(
 
     input_shape = (batch_size, 3, 640, 640)
     torch_input_tensor = torch.randn(input_shape, dtype=torch.float32)
-    tt_inputs_host = ttnn.from_torch(torch_input_tensor, dtype=ttnn.bfloat16, layout=ttnn.ROW_MAJOR_LAYOUT)
 
     inference_iter_count = 10
     inference_time_iter = []
     for iter in range(0, inference_iter_count):
         t0 = time.time()
-        output = yolov8s_trace_2cq.execute_yolov8s_trace_2cqs_inference(tt_inputs_host)
+        output = yolov8s_trace_2cq.run(torch_input_tensor)
         t1 = time.time()
         inference_time_iter.append(t1 - t0)
     yolov8s_trace_2cq.release_yolov8s_trace_2cqs_inference()
