+++ models/demos/yolov8s/tests/yolov8s_performant.py
@@ -3,7 +3,7 @@
 # SPDX-License-Identifier: Apache-2.0
 
 import ttnn
-from models.experimental.yolov8s.tests.yolov8s_test_infra import create_test_infra
+from models.demos.yolov8s.tests.yolov8s_test_infra import create_test_infra
 
 try:
     from tracy import signpost
